makefile: bin depends on lib
Building the binaries depends on the htx64 libraries but the makes
files don't currently reflect this.  So doing a parallel build gives
errors like:

  gcc   -L/home/mikey/src/HTX/export/ppc64le//lib// nf.o start_msg.o shell.o screens.o mmenu.o misc.o hxssup.o help.o search_win.o hang_monitor.o get_dispid.o erpt.o edit.o disp_dst.o build_ipc.o T_device.o R_device.o IT_device.o COE_devic.o A_device.o ART_device.o AH_system.o AH_device.o random_ahd.o equaliser.o hotplug.o -Wl,--start-group -lrt -lhtx64 -lpthread -lc -lcfgc64 -lncurses -lc -Wl,--end-group -o hxssup
  /usr/bin/ld: cannot find -lhtx64

Rerunning the build normally fixes it but this sucks, so lets fix it
properly.

The below add a dependency of lib on bin.  Now the build consistently
works first time:

  % git clean -d -f -x
  % make -j 48  &>  /dev/null
  % echo $?
  0
  %

Signed-off-by: Michael Neuling <mikey@neuling.org>