Enable travis-ci
We just do a build test for now on ubuntu 16.04

Due to ldconfig and cross compiling, we do manual symlinks for libncurses.
This is bloody awful and I should be ashamed of myself. Patches welcome.

We build inside a Docker container so that we have a programatic way
of expressing build dependencies

Signed-off-by: Stewart Smith <stewart@linux.vnet.ibm.com>